security/manager/ssl/tests/mochitest/bugs/test_bug484111.html

<html>
<head>
  <title>Test bug 484111</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="onWindowLoad()">

<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

function badCertListener() 
{
}

badCertListener.prototype = {
  badCertCaught: false,

  getInterface: function (aIID) {
    return this.QueryInterface(aIID);
  },

  QueryInterface: function(aIID) {
    if (aIID.equals(SpecialPowers.Ci.nsIBadCertListener2) ||
        aIID.equals(SpecialPowers.Ci.nsIInterfaceRequestor) ||
        aIID.equals(SpecialPowers.Ci.nsISupports))
      return this;

    throw SpecialPowers.Cr.NS_ERROR_NO_INTERFACE;
  },  

  notifyCertProblem: function(socketInfo, sslStatus, targetHost) {
    this.badCertCaught = true;
    return true;
  }
}

function onFrameLoad()
{
  ok(false, "Attackers page failed to load");
}

function onWindowLoad()
{
  var req = new XMLHttpRequest();
  var certListener = new badCertListener();
  certListener = SpecialPowers.wrapCallbackObject(certListener);
  try
  {
    req.open("GET", "https://www.bank2.com/", false);
    SpecialPowers.wrap(req).channel.notificationCallbacks = certListener;
    req.send(null);
  }
  catch(ex)
  {
    // ignore
  }
  
  ok(certListener.badCertCaught, "We Caught the invalid certificate");
  
  SimpleTest.finish();
}

</script>

<iframe src="https://www.bank2.com/" onload="onFrameLoad()"></iframe>
</body>
</html>
